// src/interfaces/user.interface.ts
export interface IUser {
    id: number;
    first_name: string;
    last_name: string;
    email: string;
    password: string;
    gender: string;
    phone: string;
    dob: string;
}

export interface IUserData {
    id: number;
    firstName: string;
    lastName: string;
    email: string;
    gender: string; 
    phone: string;
    dob: string;
}